+++ src/main/java/kitchenpos/menu/domain/MenuRepository.java
@@ -1,6 +1,9 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.List;
+
 public interface MenuRepository extends JpaRepository<Menu, Long> {
+    long countByIdIn(List<Long> ids);
 }
