+++ src/main/java/kitchenpos/menu/dto/MenuProductResponse.java
@@ -1,10 +1,19 @@
-package kitchenpos.dto;
+package kitchenpos.menu.dto;
 
-public class MenuProductRequest {
+import kitchenpos.menu.domain.MenuProduct;
+
+public class MenuProductResponse {
     private Long seq;
     private Long menuId;
     private Long productId;
     private long quantity;
+
+    public MenuProductResponse(Long seq, Long menuId, Long productId, long quantity) {
+        this.seq = seq;
+        this.menuId = menuId;
+        this.productId = productId;
+        this.quantity = quantity;
+    }
 
     public Long getSeq() {
         return seq;
@@ -36,5 +45,9 @@ public class MenuProductRequest {
 
     public void setQuantity(final long quantity) {
         this.quantity = quantity;
+    }
+
+    public static MenuProductResponse of(MenuProduct menuProduct) {
+        return new MenuProductResponse(menuProduct.getSeq(), menuProduct.getMenuId(), menuProduct.getProductId(), menuProduct.getQuantity());
     }
 }
