+++ src/main/java/kitchenpos/menu/dto/MenuResponse.java
@@ -1,6 +1,6 @@
-package kitchenpos.dto;
+package kitchenpos.menu.dto;
 
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 
 import java.math.BigDecimal;
 import java.util.List;
